+++ src/pip/_internal/distributions/installed.py
@@ -1,0 +1,10 @@
+from pip._internal.distributions.base import AbstractDistribution
+
+
+class InstalledDistribution(AbstractDistribution):
+
+    def dist(self):
+        return self.req.satisfied_by
+
+    def prep_for_dist(self, finder, build_isolation):
+        pass
