+++ src/pip/_internal/legacy_resolve.py
@@ -36,6 +36,7 @@ if MYPY_CHECK_RUNNING:
     from pip._vendor import pkg_resources
 
     from pip._internal.cache import WheelCache
+    from pip._internal.distributions import AbstractDistribution
     from pip._internal.download import PipSession
     from pip._internal.index import PackageFinder
     from pip._internal.operations.prepare import RequirementPreparer
@@ -276,7 +277,7 @@ class Resolver(object):
         return None
 
     def _get_abstract_dist_for(self, req):
-        # type: (InstallRequirement) -> DistAbstraction
+        # type: (InstallRequirement) -> AbstractDistribution
         """Takes a InstallRequirement and returns a single AbstractDist \
         representing a prepared variant of the same.
         """
